Move MessageListAction, MessageMode 👉🏻 ui-common
