+++ stream-chat-android-ui-common/api/stream-chat-android-ui-common.api
@@ -196,6 +196,62 @@ public abstract class com/getstream/sdk/chat/navigation/destinations/ChatDestina
 	protected final fun startForResult (Landroid/content/Intent;I)V
 }
 
+public final class com/getstream/sdk/chat/state/Copy : com/getstream/sdk/chat/state/MessageAction {
+	public fun <init> (Lio/getstream/chat/android/client/models/Message;)V
+}
+
+public final class com/getstream/sdk/chat/state/CustomAction : com/getstream/sdk/chat/state/MessageAction {
+	public fun <init> (Lio/getstream/chat/android/client/models/Message;Ljava/util/Map;)V
+	public synthetic fun <init> (Lio/getstream/chat/android/client/models/Message;Ljava/util/Map;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public final fun getExtraProperties ()Ljava/util/Map;
+}
+
+public final class com/getstream/sdk/chat/state/Delete : com/getstream/sdk/chat/state/MessageAction {
+	public fun <init> (Lio/getstream/chat/android/client/models/Message;)V
+}
+
+public final class com/getstream/sdk/chat/state/Edit : com/getstream/sdk/chat/state/MessageAction {
+	public fun <init> (Lio/getstream/chat/android/client/models/Message;)V
+}
+
+public final class com/getstream/sdk/chat/state/Flag : com/getstream/sdk/chat/state/MessageAction {
+	public fun <init> (Lio/getstream/chat/android/client/models/Message;)V
+}
+
+public abstract class com/getstream/sdk/chat/state/MessageAction {
+	public synthetic fun <init> (Lio/getstream/chat/android/client/models/Message;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public final fun getMessage ()Lio/getstream/chat/android/client/models/Message;
+}
+
+public abstract class com/getstream/sdk/chat/state/MessageMode {
+}
+
+public final class com/getstream/sdk/chat/state/MuteUser : com/getstream/sdk/chat/state/MessageAction {
+	public fun <init> (Lio/getstream/chat/android/client/models/Message;)V
+}
+
+public final class com/getstream/sdk/chat/state/Normal : com/getstream/sdk/chat/state/MessageMode {
+	public static final field INSTANCE Lcom/getstream/sdk/chat/state/Normal;
+}
+
+public final class com/getstream/sdk/chat/state/React : com/getstream/sdk/chat/state/MessageAction {
+	public fun <init> (Lio/getstream/chat/android/client/models/Reaction;Lio/getstream/chat/android/client/models/Message;)V
+	public final fun getReaction ()Lio/getstream/chat/android/client/models/Reaction;
+}
+
+public final class com/getstream/sdk/chat/state/Reply : com/getstream/sdk/chat/state/MessageAction {
+	public fun <init> (Lio/getstream/chat/android/client/models/Message;)V
+}
+
+public final class com/getstream/sdk/chat/state/Thread : com/getstream/sdk/chat/state/MessageMode {
+	public fun <init> (Lio/getstream/chat/android/client/models/Message;)V
+	public final fun getParentMessage ()Lio/getstream/chat/android/client/models/Message;
+}
+
+public final class com/getstream/sdk/chat/state/ThreadReply : com/getstream/sdk/chat/state/MessageAction {
+	public fun <init> (Lio/getstream/chat/android/client/models/Message;)V
+}
+
 public abstract interface class com/getstream/sdk/chat/utils/DateFormatter {
 	public static final field Companion Lcom/getstream/sdk/chat/utils/DateFormatter$Companion;
 	public abstract fun formatDate (Lorg/threeten/bp/LocalDateTime;)Ljava/lang/String;
@@ -370,6 +426,23 @@ public abstract class com/getstream/sdk/chat/viewmodel/BaseMessageListHeaderView
 	public final fun getTypingUsers ()Landroidx/lifecycle/LiveData;
 	public final fun resetThread ()V
 	public final fun setActiveThread (Lio/getstream/chat/android/client/models/Message;)V
+}
+
+public class com/getstream/sdk/chat/viewmodel/MessageComposerViewModel : androidx/lifecycle/ViewModel {
+	public fun <init> (Lio/getstream/chat/android/client/ChatClient;Lio/getstream/chat/android/offline/ChatDomain;Ljava/lang/String;)V
+	public final fun addSelectedAttachments (Ljava/util/List;)V
+	public final fun buildNewMessage (Ljava/lang/String;Ljava/util/List;)Lio/getstream/chat/android/client/models/Message;
+	public static synthetic fun buildNewMessage$default (Lcom/getstream/sdk/chat/viewmodel/MessageComposerViewModel;Ljava/lang/String;Ljava/util/List;ILjava/lang/Object;)Lio/getstream/chat/android/client/models/Message;
+	public final fun dismissMessageActions ()V
+	public final fun getActiveAction ()Lcom/getstream/sdk/chat/state/MessageAction;
+	public final fun getInput ()Lkotlinx/coroutines/flow/MutableStateFlow;
+	public final fun getSelectedAttachments ()Lkotlinx/coroutines/flow/MutableStateFlow;
+	public final fun leaveThread ()V
+	public final fun performMessageAction (Lcom/getstream/sdk/chat/state/MessageAction;)V
+	public final fun removeSelectedAttachment (Lio/getstream/chat/android/client/models/Attachment;)V
+	public final fun sendMessage (Lio/getstream/chat/android/client/models/Message;)V
+	public final fun setMessageInput (Ljava/lang/String;)V
+	public final fun setMessageMode (Lcom/getstream/sdk/chat/state/MessageMode;)V
 }
 
 public final class com/getstream/sdk/chat/viewmodel/MessageInputViewModel : androidx/lifecycle/ViewModel {
